requirements.txt
# Step 1: PDF parsing
pydantic>=2.0.0
unstructured[pdf]>=0.10.0

# Step 2: Boundary-aware chunking (token-based, sentence splitting)
tiktoken>=0.5.0
nltk>=3.8.0
pyyaml>=6.0.0

# Step 4: Embeddings + vectorstore
numpy>=1.24.0

# Step 5: Synthetic QA generation (LLM + optional embeddings)
openai>=1.0.0

# Tests
pytest>=7.0.0
